Move League service providers into the League namespace
